requirements.txt
requests==2.31.0
gitpython==3.1.40
rich==13.7.0